I want a Pokémon like game with three.js
I want to make it look top bottom like Pokémon red game, but allow to zoom in and out.

To begin with let's start with having placeholder shapes for all the items in the game. I will provide actual assets later on.

So for now let's have the main hero walking around huge map. there are houses and different landscapes. For now this is it