the top and bottom buttons on mobile joystick are reversed. Please fix it.
Also I only see one type of ground. I want to add another type. It will be like in Pokémon red. After walking on it for a few seconds, let's say 2 to 10, but measure it by distance walked not time. Then randomly you get attacked and enter fight mode. The fight will be like heroes 3. Tour team from left and enemy from the right. It is a turn base fight. For now let's just give your hero the ability to move few cells or shout in a fight. And the enemy will only shoot. Show some basic projectiles when so.eone shoots