build a logic game where the user need to count blocks. Use three.js for it.
Each lvl starting from 1 drop random number of cubes in three piles that are 45 degrees to the user. Position them in such that two are in front and one in the back. Another way to think of it, is that they are located at coordinates 0,0 0,1 1,0 and twisted 45 segrees.

the 0,0 must be the tallest of the three. Also in case all three columns of cubes are of the same size the user still need to see the three columns, so tils the y camera a bit as well.

Give the user 4 options to choose the answer from each round. Make them sequential, positive, numbers.